model update

+++ core/src/main/java/kr/co/bootpay/model/request/Item.java
@@ -4,7 +4,7 @@ public class Item {
     public String name; // 상품명
     public int qty; // 수량
     public String id; // 상품 고유키
-    public long price; // 상품단가
+    public double price; // 상품단가
     public String cat1; // 카테고리 상
     public String cat2; // 카테고리 중
     public String cat3; // 카테고리 하
